requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Bcrypt>=1.0
Flask-Mail>=0.9.1
Flask-Login>=0.6
Flask-Migrate>=4.0
python-dotenv>=1.0
requests>=2.28
reportlab>=4.0
SQLAlchemy>=1.4
Pillow>=9.0
Flask-WTF>=1.0